requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-openai
ultralytics
matplotlib
Pillow
langchain-core
sqlalchemy
psycopg2-binary
alembic